feat: add workaround for wine's unstable dinput instance guids
If the scripts detect that they are running inside a wine environment,
they now create configuration files not only for the current instance
GUID of the virtual joystick device, but also for at most 5 counter
steps below and 5 counter steps above.

Context:

Wine calculates the first part of the instance GUID as follows:

instance->guidInstance.Data1 ^= handle;

The remainder of the GUID is static.

The `handle` is not stable as it is based on a counter which gets
incremented whenever a new input device is allocated:

static UINT32 alloc_rawinput_handle(void)
{
    static LONG counter = WINE_KEYBOARD_HANDLE + 1;
    return InterlockedIncrement(&counter);
}